# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/graphs": {
      "get": {
        "operationId": "GraphController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/GraphSummaryDto"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List graphs the caller can see",
        "tags": [
          "graphs"
        ]
      },
      "post": {
        "operationId": "GraphController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGraphDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/GraphDetailDto"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "A graph with that id already exists.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "error"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    },
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "VALIDATION_ERROR"
                        },
                        "message": {
                          "type": "string",
                          "example": "Request validation failed."
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": true,
                          "description": "Present only when the error carries structured context."
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create an empty graph",
        "tags": [
          "graphs"
        ]
      }
    },
    "/api/graphs/{graphId}": {
      "get": {
        "description": "A graph the caller may not see answers 404, never 403.",
        "operationId": "GraphController_findOne",
        "parameters": [
          {
            "name": "graphId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/GraphDetailDto"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No such graph, or not visible to the caller.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "error"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    },
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "VALIDATION_ERROR"
                        },
                        "message": {
                          "type": "string",
                          "example": "Request validation failed."
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": true,
                          "description": "Present only when the error carries structured context."
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Read a graph with its nodes and edges",
        "tags": [
          "graphs"
        ]
      },
      "delete": {
        "operationId": "GraphController_remove",
        "parameters": [
          {
            "name": "graphId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/GraphSummaryDto"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No such graph.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "error"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    },
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "VALIDATION_ERROR"
                        },
                        "message": {
                          "type": "string",
                          "example": "Request validation failed."
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": true,
                          "description": "Present only when the error carries structured context."
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a graph, its content, and its events",
        "tags": [
          "graphs"
        ]
      }
    },
    "/api/graphs/{graphId}/content": {
      "put": {
        "description": "Atomic. Invariants are checked before the transaction opens, so an invalid edge costs nothing and names itself.",
        "operationId": "GraphController_replaceContent",
        "parameters": [
          {
            "name": "graphId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplaceGraphContentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/GraphDetailDto"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "The caller can see the graph but does not own it.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "error"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    },
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "VALIDATION_ERROR"
                        },
                        "message": {
                          "type": "string",
                          "example": "Request validation failed."
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": true,
                          "description": "Present only when the error carries structured context."
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "The structure changed since it was read.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "error"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    },
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "VALIDATION_ERROR"
                        },
                        "message": {
                          "type": "string",
                          "example": "Request validation failed."
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": true,
                          "description": "Present only when the error carries structured context."
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "A dangling endpoint, a self-loop, or a duplicate edge.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "error"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    },
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "VALIDATION_ERROR"
                        },
                        "message": {
                          "type": "string",
                          "example": "Request validation failed."
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": true,
                          "description": "Present only when the error carries structured context."
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Replace every node and edge",
        "tags": [
          "graphs"
        ]
      }
    },
    "/api/graphs/{graphId}/topology/snapshot": {
      "get": {
        "description": "Read directly from runtime rows, not folded from events. `revision` is the sequence the snapshot reflects — subscribe from it.",
        "operationId": "GraphController_snapshot",
        "parameters": [
          {
            "name": "graphId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/TopologySnapshotDto"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No such graph.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "error"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    },
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "VALIDATION_ERROR"
                        },
                        "message": {
                          "type": "string",
                          "example": "Request validation failed."
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": true,
                          "description": "Present only when the error carries structured context."
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Current runtime state",
        "tags": [
          "graphs"
        ]
      }
    },
    "/api/graphs/{graphId}/topology/resync": {
      "get": {
        "description": "The HTTP counterpart of the WebSocket replay. Answers `resync` when the gap is older than the retained window, which means the client must take a fresh snapshot rather than assume continuity.",
        "operationId": "GraphController_resync",
        "parameters": [
          {
            "name": "graphId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "lastSequence",
            "required": true,
            "in": "query",
            "description": "The highest sequence the client already applied.",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/TopologyReplayDto"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No such graph.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "error"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    },
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "VALIDATION_ERROR"
                        },
                        "message": {
                          "type": "string",
                          "example": "Request validation failed."
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": true,
                          "description": "Present only when the error carries structured context."
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Events missed since a sequence",
        "tags": [
          "graphs"
        ]
      }
    },
    "/api/graphs/{graphId}/topology/events": {
      "post": {
        "description": "Allocates a sequence, stores the event, updates runtime state, and fans it out — the first three in one transaction. Exists so a monitoring agent can feed the stream; the UI only reads.",
        "operationId": "GraphController_publish",
        "parameters": [
          {
            "name": "graphId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishTopologyEventDto"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/TopologyReplayDto"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No such graph.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "error"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    },
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "VALIDATION_ERROR"
                        },
                        "message": {
                          "type": "string",
                          "example": "Request validation failed."
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": true,
                          "description": "Present only when the error carries structured context."
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Publish a runtime event",
        "tags": [
          "graphs"
        ]
      }
    },
...
  },
  "info": {
    "title": "React Boilerplate API",
    "description": "Backend for the React boilerplate. Every response uses the shared envelope: `{ success: true, data }` on success and `{ success: false, error: { code, message, details? } }` on failure. HTTP status describes the transport outcome; `error.code` describes the domain outcome.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "rb_refresh": {
        "type": "apiKey",
        "in": "cookie",
        "name": "rb_refresh"
      }
    },
    "schemas": {
...
      "GraphSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "ownerId": {
            "type": "string"
          },
          "visibility": {
            "type": "string",
            "enum": [
              "private",
              "shared"
            ]
          },
          "version": {
            "type": "number",
            "description": "Optimistic lock on the structure."
          },
          "sequence": {
            "type": "number",
            "description": "Highest runtime event sequence. Unrelated to version."
          },
          "nodeCount": {
            "type": "number"
          },
          "edgeCount": {
            "type": "number"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "ownerId",
          "visibility",
          "version",
          "sequence",
          "nodeCount",
          "edgeCount",
          "updatedAt"
        ]
      },
...
      "CreateGraphDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "visibility": {
            "type": "string",
            "enum": [
              "private",
              "shared"
            ],
            "default": "private"
          }
        },
        "required": [
          "id",
          "title"
        ]
      },
      "ReplaceGraphContentDto": {
        "type": "object",
        "properties": {
          "expectedVersion": {
            "type": "number",
            "description": "The structure version last read. A mismatch answers 409."
          },
          "nodes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GraphNodeDto"
            }
          },
          "edges": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GraphEdgeDto"
            }
          }
        },
        "required": [
          "expectedVersion",
          "nodes",
          "edges"
        ]
      },
      "TopologySnapshotDto": {
        "type": "object",
        "properties": {
          "topologyId": {
            "type": "string"
          },
          "revision": {
            "type": "number",
            "description": "The sequence this snapshot reflects. Subscribe from here."
          },
          "capturedAt": {
            "type": "number"
          },
          "nodes": {
            "type": "object"
          },
          "edges": {
            "type": "object"
          }
        },
        "required": [
          "topologyId",
          "revision",
          "capturedAt",
          "nodes",
          "edges"
        ]
      },
...
    }
  }
}